Fix Dapper type mismatch in NicknameResolverService: update `ChatUser.MessageCount` from `int` to `long` for `bigint` compatibility. Update CHANGELOG.

+++ WatchmenBot/Features/Search/Services/NicknameResolverService.cs
@@ -38,7 +38,7 @@ public class NicknameResolverService(
     /// <summary>
     /// Chat user info for nickname resolution
     /// </summary>
-    public record ChatUser(string DisplayName, string? Username, int MessageCount);
+    public record ChatUser(string DisplayName, string? Username, long MessageCount);
 
     /// <summary>
     /// Result of nickname resolution
